Disable Graphite even if isl is present (see #1)
If isl is installed, the Graphite optimizations are built. They are
optional. I have found out that if ISL is installed via Homebrew, it
fails to build unless I set the --without-isl flag on the Formula.

I suspect this is happening because Homebrew installs a newer version
and I'm attempting to build an older GCC version, but I'd rather keep
using old versions for now until I finish porting the old buildscript
and check that everything works the same.